Moved some things around for spending_by_category
